Move sensors to sensor.py
