+++ custom_components/frank_energie/const.py
@@ -1,18 +1,5 @@
 from __future__ import annotations
 
-from dataclasses import dataclass
-from typing import Callable
-
-from homeassistant.components.sensor import SensorEntityDescription
-from homeassistant.const import (
-    CURRENCY_EURO,
-    ENERGY_KILO_WATT_HOUR,
-    VOLUME_CUBIC_METERS,
-)
-from homeassistant.helpers.typing import StateType
-
-from .price_data import PriceData
-
 ATTRIBUTION = "Data provided by Frank Energie"
 DOMAIN = "frank_energie"
 DATA_URL = "https://frank-graphql-prod.graphcdn.app/"
@@ -25,132 +12,3 @@ ATTR_TIME = "from_time"
 
 DATA_ELECTRICITY = 'electricity'
 DATA_GAS = 'gas'
-
-
-@dataclass
-class FrankEnergieEntityDescription(SensorEntityDescription):
-    """Describes Frank Energie sensor entity."""
-    value_fn: Callable[[dict[PriceData]], StateType] = None
-    attr_fn: Callable[[dict[PriceData]], dict[str, StateType | list]] = lambda _: {}
-
-
-SENSOR_TYPES: tuple[FrankEnergieEntityDescription, ...] = (
-    FrankEnergieEntityDescription(
-        key="elec_markup",
-        name="Current electricity price (All-in)",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.total,
-        attr_fn=lambda data: {'prices': data[DATA_ELECTRICITY].asdict('total')},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_market",
-        name="Current electricity market price",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.market_price,
-        attr_fn=lambda data: {'prices': data[DATA_ELECTRICITY].asdict('market_price')},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_tax",
-        name="Current electricity price including tax",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.market_price_with_tax,
-        attr_fn=lambda data: {'prices': data[DATA_ELECTRICITY].asdict('market_price_with_tax')},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_tax_vat",
-        name="Current electricity VAT price",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.market_price_tax,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_sourcing",
-        name="Current electricity sourcing markup",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.sourcing_markup_price,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_tax_only",
-        name="Current electricity tax only",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].current_hour.energy_tax_price,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_markup",
-        name="Current gas price (All-in)",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.total,
-        attr_fn=lambda data: {'prices': data[DATA_GAS].asdict('total')},
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_market",
-        name="Current gas market price",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.market_price,
-        attr_fn=lambda data: {'prices': data[DATA_GAS].asdict('market_price')},
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_tax",
-        name="Current gas price including tax",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.market_price_with_tax,
-        attr_fn=lambda data: {'prices': data[DATA_GAS].asdict('market_price_with_tax')},
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_tax_vat",
-        name="Current gas VAT price",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.market_price_tax,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_sourcing",
-        name="Current gas sourcing price",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.sourcing_markup_price,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_tax_only",
-        name="Current gas tax only",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].current_hour.energy_tax_price,
-        entity_registry_enabled_default=False,
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_min",
-        name="Lowest gas price today",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].today_min.total,
-        attr_fn=lambda data: {ATTR_TIME: data[DATA_GAS].today_min.date_from},
-    ),
-    FrankEnergieEntityDescription(
-        key="gas_max",
-        name="Highest gas price today",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{VOLUME_CUBIC_METERS}",
-        value_fn=lambda data: data[DATA_GAS].today_max.total,
-        attr_fn=lambda data: {ATTR_TIME: data[DATA_GAS].today_max.date_from},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_min",
-        name="Lowest energy price today",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].today_min.total,
-        attr_fn=lambda data: {ATTR_TIME: data[DATA_ELECTRICITY].today_min.date_from},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_max",
-        name="Highest energy price today",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].today_max.total,
-        attr_fn=lambda data: {ATTR_TIME: data[DATA_ELECTRICITY].today_max.date_from},
-    ),
-    FrankEnergieEntityDescription(
-        key="elec_avg",
-        name="Average electricity price today",
-        native_unit_of_measurement=f"{CURRENCY_EURO}/{ENERGY_KILO_WATT_HOUR}",
-        value_fn=lambda data: data[DATA_ELECTRICITY].today_avg,
-    ),
-)
